- Wouter Wijngaards fixed signed/unsigned warning in   csLoader::CreateSequenceParameters.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@20676 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/csparser/prsseq.cpp
@@ -219,7 +219,7 @@ csPtr<iEngineSequenceParameters> csLoader::CreateSequenceParameters (
 {
   error = false;
   csRef<iEngineSequenceParameters> params = sequence->CreateParameterBlock ();
-  int found_params = 0;
+  size_t found_params = 0;
   csRef<iDocumentNodeIterator> it = node->GetNodes ();
   while (it->HasNext ())
   {
